% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{rothko_palette}
\alias{rothko_palette}
\title{A Wes Anderson palette generator}
\usage{
rothko_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{Magenta, Black, Green on Orange}, \code{Untitled 1949},  \code{Number 5},
\code{Violet, Black, Orange, Yellow on White and Red},
\code{Violet, Black, Orange, Yellow on White and Red2},
\code{White Center}, \code{Blue, Green, and Brown},
\code{Purple, White, and Red},  \code{Number 10},
\code{Rust and Blue} ,\code{Untitled 1953}, \code{Ochre and Red on Red},
\code{Red, Black, White on Yellow}, \code{Green and Tangerine on Red},
\code{Light Red Over Black}, \code{Light Cloud, Dark Cloud}, \code{Number 14},
\code{Orange, Red, Yellow}, \code{Number 16}, \code{Blue and Grey},
\code{Untitled 1969}}

\item{n}{Number of colors desired.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from Wes Anderson movies.
}
\examples{
rothko_palette("Untitled 1949")
rothko_palette("Number 10")
rothko_palette("Purple, White, and Red")
rothko_palette("Purple, White and Red", 3)
}
\keyword{colors}
